pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Task Manager is a task management system. It allows you to set tasks, assign executors and change their statuses."
authors = ["BogdanBarylo <barylob@gmail.com>"]
readme = "README.md"
packages = [{include = "task_manager"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.3"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
dj-database-url = "^2.1.0"
gunicorn = "^21.2.0"
whitenoise = {extras = ["brotli"], version = "^6.6.0"}
django-bootstrap5 = "^23.4"
uvicorn = "^0.28.0"
django-filter = "^24.1"
rollbar = "^0.16.3"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
coverage = "^7.4.4"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
